fineControl has been fixed
The function that let the controller be less sensitive at lower values
as been fixed and implemented into the code. This means should the
controller be moved up to a value of say 50, it will only power motor to
around 20, this way there is more precise control. However should the
controller be pushed up to its maximum of 127, it will still set motors
to 127 so no speed is sacrificed. The problem itself lay within the
floating point calculators of the C language, it somehow could not
implicitly convert 127 to 127.0 when set on a float variable.

+++ ImpulseV1/files/codebits/finecontrol.c
@@ -7,9 +7,9 @@ int fineControl(int input) //input from value, mod is set to driver preferences
 
 		input = abs(input);
 		
-		float percentage = (float)(input/127); //Find relative speed to max speed
-		float fineMod = (float)(pow(fineControlValue,percentage)-1); //find the modifier value to multiply by
-		float scalingValue = (float)(127/(fineControlValue-1)); //value to multiply back by after modification
+		float percentage = (input/127.0); //Find relative speed to max speed
+		float fineMod = (pow(fineControlValue,percentage)-1); //find the modifier value to multiply by
+		float scalingValue = (127.0/(fineControlValue-1)); //value to multiply back by after modification
 		
 		int output = (int)(round(scalingValue*sign*fineMod));
